So, the thread attr stuff is opaque according to the offical APR manual. The upstream FS devs blithely ignore this. Like turkeys.
Also, while we're at it, re-write the function to give insight
into errors incase there are errors. Note that at this time, none
of the callers actually check return values of anything so....

Anyway, moral of the story is that that this is all a giant pile of
shit; upstream should be embarrased. This stuff needs to be purged.