I was being reckless with the last commit, and hadn't inspected the memory alloc calls to see if they were simple stubs (or not).
Turns out they are not stubs, and it is innapropriate to replace
those calls with the apr equivalents.

Note that, this is only because FS is using APR to handle memory,
and because it is a heavily threaded application, the shared
memory used by the apr pool functionality is protected via mutexes.

Since I am uncertain if the underlying pool functionality is
threadsafe, I will revert to the original behaviour.

In summary... threads are a pain in the ass.